backend/requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
pillow==11.2.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
opencv-python-headless==4.11.0.86
matplotlib==3.10.3

ultralytics==8.3.108
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
requests==2.31.0
